requirements.txt
asgiref==3.7.2
Django==4.1
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-js-asset==2.2.0
djangorestframework==3.14.0
pytz==2024.1
sqlparse
